# Major perf improvement by decorating unsafe functions with AES target.