This enables inlining of the AES intrinsics and fixes the big performance difference in the C version as discussed in #1. 

`test digest_3_1m` benchmark went from 110,758 ns to 38,966 ns (8850 -> 26069 MB/s) on my MBP13, which is very nice!

Big thanks to @bitshifter for suggesting this was the cause of the perf delta.